Move api/server/router/swarm to daemon/server/router/swarm
