+++ daemon/server/router/swarm/cluster.go
