+++ daemon/server/router/swarm/cluster_routes.go
