+++ daemon/server/router/swarm/helpers.go
