+++ daemon/server/router/swarm/helpers_test.go
